Add generic ValidateToken - only functional change is that the 15s provider timeout is used for all providers.
